Move `flt2dec::{Formatted, Part}` to dedicated module
They are used by integer formatting as well and is not exclusive to float.
